Bug Fix
Bug where after the client sends an input, the server would send back the real response message followed by 3 of the "guess too low" messages. Initially thought that the server was sending all the messages in one send, however it was found that the server recv/send loop ran 4 times. This suggested that the bug was that it had to handle input using 4 loops. This meant that the client was sending 4 segments of data (each segment is 512 bytes and the original send sent 2000 bytes). Changing the bytes sent to 512 fixed this issue as the server can handle all the client's sent data with one loop and not have to re-process the input.